kubefloww-trainer/requirements.txt

pip
kfp
transformers
lightning
torchmetrics
scikit-learn
black
datasets
torchserve
torch-model-archiver
ipykernel
matplotlib
pytorch-kfp-components